fix(core): extract inner classes with annotations
