+++ spring-ai-alibaba-core/src/main/java/com/alibaba/cloud/ai/advisor/DocumentRetrievalAdvisor.java
@@ -15,11 +15,12 @@
  */
 package com.alibaba.cloud.ai.advisor;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.ai.chat.client.ChatClientRequest;
 import org.springframework.ai.chat.client.ChatClientResponse;
 import org.springframework.ai.chat.client.advisor.api.AdvisorChain;
@@ -31,7 +32,16 @@ import org.springframework.ai.rag.Query;
 import org.springframework.ai.rag.retrieval.search.DocumentRetriever;
 import org.springframework.util.Assert;
 
+/**
+ * Title Document retrieval advisor.<br>
+ * Description Document retrieval advisor.<br>
+ *
+ * @author yuanci.ytb
+ * @since 1.0.0-M2
+ */
 public class DocumentRetrievalAdvisor implements BaseAdvisor {
+
+	private static final Logger logger = LoggerFactory.getLogger(DocumentRetrievalAdvisor.class);
 
 	private static final PromptTemplate DEFAULT_PROMPT_TEMPLATE = new PromptTemplate("""
 			{query}
@@ -160,177 +170,6 @@ public class DocumentRetrievalAdvisor implements BaseAdvisor {
 			.chatResponse(chatResponseBuilder.build())
 			.context(chatClientResponse.context())
 			.build();
-	}
-
-	public static class CompositeDocumentRetriever implements DocumentRetriever {
-
-		private final List<DocumentRetriever> retrievers;
-
-		private final int maxResultsPerRetriever;
-
-		private final ResultMergeStrategy mergeStrategy;
-
-		public enum ResultMergeStrategy {
-
-			SIMPLE_MERGE,
-
-			SCORE_BASED,
-
-			ROUND_ROBIN
-
-		}
-
-		public CompositeDocumentRetriever(List<DocumentRetriever> retrievers) {
-			this(retrievers, 10, ResultMergeStrategy.SCORE_BASED);
-		}
-
-		public CompositeDocumentRetriever(List<DocumentRetriever> retrievers, int maxResultsPerRetriever) {
-			this(retrievers, maxResultsPerRetriever, ResultMergeStrategy.SCORE_BASED);
-		}
-
-		public CompositeDocumentRetriever(List<DocumentRetriever> retrievers, int maxResultsPerRetriever,
-				ResultMergeStrategy mergeStrategy) {
-			Assert.notNull(retrievers, "Retrievers list must not be null!");
-			Assert.isTrue(!retrievers.isEmpty(), "Retrievers list must not be empty!");
-			Assert.isTrue(maxResultsPerRetriever > 0, "MaxResultsPerRetriever must be positive!");
-			Assert.notNull(mergeStrategy, "MergeStrategy must not be null!");
-
-			this.retrievers = new ArrayList<>(retrievers);
-			this.maxResultsPerRetriever = maxResultsPerRetriever;
-			this.mergeStrategy = mergeStrategy;
-		}
-
-		@Override
-		public List<Document> retrieve(Query query) {
-			if (mergeStrategy == ResultMergeStrategy.ROUND_ROBIN) {
-				return roundRobinRetrieve(query);
-			}
-
-			List<Document> allDocuments = new ArrayList<>();
-
-			for (DocumentRetriever retriever : retrievers) {
-				try {
-					List<Document> documents = retriever.retrieve(query);
-					if (documents != null && !documents.isEmpty()) {
-						List<Document> limitedDocuments = documents.stream()
-							.limit(maxResultsPerRetriever)
-							.collect(Collectors.toList());
-						allDocuments.addAll(limitedDocuments);
-					}
-				}
-				catch (Exception e) {
-					System.err.println("Error retrieving from one of the retrievers: " + e.getMessage());
-				}
-			}
-
-			return mergeResults(allDocuments);
-		}
-
-		private List<Document> roundRobinRetrieve(Query query) {
-			List<List<Document>> allResults = new ArrayList<>();
-
-			for (DocumentRetriever retriever : retrievers) {
-				try {
-					List<Document> documents = retriever.retrieve(query);
-					if (documents != null && !documents.isEmpty()) {
-
-						List<Document> limitedDocuments = documents.stream()
-							.limit(maxResultsPerRetriever)
-							.collect(Collectors.toList());
-						allResults.add(limitedDocuments);
-					}
-					else {
-						allResults.add(new ArrayList<>());
-					}
-				}
-				catch (Exception e) {
-					System.err.println("Error retrieving from one of the retrievers: " + e.getMessage());
-					allResults.add(new ArrayList<>());
-				}
-			}
-
-			List<Document> result = new ArrayList<>();
-			int maxSize = allResults.stream().mapToInt(List::size).max().orElse(0);
-
-			for (int i = 0; i < maxSize; i++) {
-				for (List<Document> documents : allResults) {
-					if (i < documents.size()) {
-						result.add(documents.get(i));
-					}
-				}
-			}
-
-			return result;
-		}
-
-		private List<Document> mergeResults(List<Document> documents) {
-			if (documents.isEmpty()) {
-				return documents;
-			}
-
-			switch (mergeStrategy) {
-				case SIMPLE_MERGE:
-					return documents;
-				case SCORE_BASED:
-					return documents.stream().sorted((d1, d2) -> {
-						Double score1 = d1.getScore();
-						Double score2 = d2.getScore();
-
-						if (score1 == null)
-							score1 = 0.0;
-						if (score2 == null)
-							score2 = 0.0;
-						return Double.compare(score2, score1);
-					}).collect(Collectors.toList());
-				case ROUND_ROBIN:
-					return documents;
-				default:
-					return documents;
-			}
-		}
-
-		public static class Builder {
-
-			private List<DocumentRetriever> retrievers = new ArrayList<>();
-
-			private int maxResultsPerRetriever = 10;
-
-			private ResultMergeStrategy mergeStrategy = ResultMergeStrategy.SCORE_BASED;
-
-			public Builder addRetriever(DocumentRetriever retriever) {
-				if (retriever != null) {
-					this.retrievers.add(retriever);
-				}
-				return this;
-			}
-
-			public Builder retrievers(List<DocumentRetriever> retrievers) {
-				if (retrievers != null) {
-					this.retrievers.addAll(retrievers);
-				}
-				return this;
-			}
-
-			public Builder maxResultsPerRetriever(int maxResultsPerRetriever) {
-				this.maxResultsPerRetriever = maxResultsPerRetriever;
-				return this;
-			}
-
-			public Builder mergeStrategy(ResultMergeStrategy mergeStrategy) {
-				this.mergeStrategy = mergeStrategy;
-				return this;
-			}
-
-			public CompositeDocumentRetriever build() {
-				return new CompositeDocumentRetriever(retrievers, maxResultsPerRetriever, mergeStrategy);
-			}
-
-		}
-
-		public static Builder builder() {
-			return new Builder();
-		}
-
 	}
 
 }
